Fix fields and allow them to be extended
The fields part of the datastore_search's response is used by reclineJS. It is
critical, therefore, that it is output in the correct format and indeed before this
commit it was not (I had changed the 'id' key in each field definition to 'name').
That is fixed in this commit and additionally the fields return value is now
extensible through the plugin framework.