Implement an autocomplete action
The autocomplete action can be used to return values from
a specific field given a search term. The search term is
used in a prefix on the field. This commit has a few small
refactors in it to share the search code between the existing
datastore_search action and the new datastore_autocomplete
action.