Reindex should reread everything from mongo
Reindexing should reread everything from mongo and put it
into the index to really be useful. This has been specifically
implemented like this with maps in mind so that the user can
change/set the lat/long fields and then either manually or
automatically trigger a reindex to get the data into the
meta.geo field.