Add run_query parameter to datastore_search to retrieval of raw query
This has been added with specific use cases in mind, namely for map tile
servers using the elasticsearch backend to generate tile images. In this
scenario we want to run the query and generate the tile on a different
server but still use the same logic to determine the query from the
selected user parameters. To avoid the horrible scenario where the tile
server and the core ckan extension have duplicated code it makes more
sense to allow the datastore_search action to return the query it would
have run instead of the result. This can then be used by the tile server
to retrieve the necessary data it needs. The other key advantage of this
process is that we can also then loop through the implementing plugins
as well.