Add geo filtering to the datastore search API
A filter with the key __geo__ can now be passed through in a search to
add geographic filters to the query. The value associated should contain
a GeoJSON geometry object containing a type key (can currently have
values of "Point", "Polygon" or "Multipolygon") and a coordinates key
with a list of coordinates (the exact structure of this varies between
the types). If the "Point" type is used then an additional, non-GeoJSON
key must be passed - "distance", allowing queries within a circle around
a point.