Change how stats are stored during ingest/index
The stats model has been altered to store ingest and index operation
stats separately. This is because not all ingest and index operations
will occur in tandem, in fact it's likely there will be a bunch of index
operations that don't involve and ingest.

This has knock on effects throughout the importing code as we need to
create and update the stats records differently. There are also changes
in the retrieval and rendering of these operations to the user on the
datastore page.