Improve handling of replaced records
The default functionality when users upload a file is that the contents
of that file replaces the existing data in the system already. For the
versioning backend this means we need to have the ability to remove
records if they don't appear in the new file. This was achieved
previously by always storing a diff on each record, regardless of
whether if had changed. This undermines the whole diffing system though
as it means we're storing loads of potentially empty diffs. This has now
been changed so that we use a temporary collection to keep track of the
records that were in the new file but weren't changed, allowing us to
"delete" the records that weren't in the file by examining a combination
of the ones that have a latest version in the resource collection and
those that weren't in the list of unchanged records.

Note that the decision to use a temporary collection is purely so that
we can support large resources without running our of memory by creating
and enormous set of ids.