Refactor queueing so that we use objects to pass args
This avoids passing loads of arguments as they are nicely encapsulated
together. It also avoids an issue I came across in testing where I was
using the API to upsert a large number of records which resulted in the
entire list of records being printed out in the rq logs because it
prints the separate arguments. By passing an object we can control the
repr/str of the single arg and show some useful info.