Use a more robust encoding detection mechanism
For *sv files we now take a 400 line sample of the data and attempt to
detect the encoding using chardet. This technique is also used in
requests itself (though it reads the whole source URL which we can't do)
and in ckan's messytables replacement lib tabulator. We could use the
tabulator lib, and we should really, but right now we can't due to
compatibility issues - we're on CKAN 2.3.x which requires
sqlalchemy==0.7.8, whereas tabulator requires sqlalchemy>=0.9.6,<2.0.