Use aliases to control access to data in elasticsearch
This isn't technically necessary yet, but it will be when we do
cross-resource search. This is because right now we're controlling
access through the CKAN code and there's no way of getting to the data
without going through that code. In the future we'll be allowing
searches which have dynamic indices defined (i.e. elasticsearch will
resolve them, not us in the CKAN code) and therefore we need to have
information about privacy there too.

The structure put in place in this commit is to use the default index
as the private version of the data, which all resources in the datastore
get (nhm-<resource_id>). And then if the resource is public we create an
alias to the private index (pubnhm-<resource_id>).