Update resource deletion code to maintain old versions
Previously, we were just removing all data from mongo and elasticsearch
but this prevented the data in old versions from being seen which kinda
defeats the point of having a fully versioned datastore. This commit
alters that by correctly just deleting the current data. This means that
if you attempt to get any data for the resource as it looks at the most
recent version, you'll see nothing, but if you look for an older version
you get the data.