Refactor the notify usage to be a bit more predictable
Say bye bye to IResourceUrlChange, it just confuses things. The notify
function is pretty dodgy but this cleans up the usage of it a little bit
to make it easier to understand. This commit also just tidies up a
couple of `inherit=True`s that didn't need to be in use.