Respect the current version's column order before other old versions
In commit b9b64911c2c0f5ecc4c3e0c77c1e009e3c9095a2 we added code to
handle a specific ordering problem. The problem was:

---
in v1 a single record is added:
  {'x': -5, 't': 'hi', 'a': true, 'b': 29}

in v2 another single record is added and the existing record is kept:
  {'e': false}

therefore the fields for v1 are: ['x', 't', 'a', 'b']
and for v2 the fields are: ['x', 't', 'a', 'b', 'e']

essentially, the order of the original v1 fields is maintained with the
addition of a new field in v2.
---

This works great for partial updates, but not so well if:

in v1 a single record is added:
  {'x': -5, 't': 'hi', 'a': true, 'b': 29}

in v2 another single record is added but the existing record is
replaced:
  {'b': 29, 'a': true, 't': 'hi', 'x': -5}

we'd expect v1's fields to look like this: ['x', 't', 'a', 'b']
and v2's fields to look like this: ['b', 'a', 't', 'x']. But they'd
actually be the same because v1's field order is used before v2s. This
commit fixes this problem by reversing the order so that we use the
order from the version we're querying first and then work backwards,
adding the fields from previous versions as we go. This fixes this
problem and will result in ['e', 'x', 't', 'a', 'b'] for v2 in the first
problem outline - which is perfectly acceptable I think.