Use the revision timestamp instead of last modified
Turns out last_modified is only updated during a new file upload* which
makes sense once you learn that the last_modified attribute on resources
is meant to indicate when the *data* was last updated, not the metadata.
For the metadata the last_modified attribute isn't updated but the
revision timestamp is. Luckily the revision timestamp is also updated
when a new file is uploaded so we can use that for everything.

See this issue for info on the last_modified/revision timestamp
difference: https://github.com/ckan/ckan/issues/1874.

*this doesn't apply if you change the url, only if you upload a file.