Resolve issue with universal lines
Typically, newlines on windows and unix systems are different - \r\n and
\n respectively. In python2 this is handleable by passing rU when
opening the file and python will switch the \r\n bytes with \n before
giving you the data. We need to able to read all kinds of line endings
so we need to do this, however, we're passing a file object allowing
access to the resource's data in rb mode and there is no way to wrap the
stream and do universal lines in the standard lib*. This commit contains
a new function which does exactly this, given an open rb file we read
from it and replace the weird line endings. This is only needed in the
SVReader.

*we could do `open(fp.name, 'rU')`, however, that puts reliance on the
fp variable allowing this and then working when you do it which couples
these two areas too tightly in my opinion (in the code right now it is
already possible for that fp to be a gzipped file or just a normal
uncompressed file). We could use the reader's compressible property to
determine how to reopen the file but this just starts to get a bit messy
when you actually try to implement it. Believe it or not, this is
easier.