Detect and error out when we find invalid unicode characters
If someone either:

- provides a file with an encoding that we can't determine
- provides a file with control characters/non-printable characters in it

we can end up writing these out into our intermediate file. This works
fine and the file will pass the prep phase but then when ingesting it
will fail as we can't read the serialised JSON data. This catches that
problem at the prep phase. It's possible that this will be a bit more
draconian than I anticipate at this point, mainly because I haven't
tested this on all files ever, however, it's probably better that we
work on that end of the spectrum than the other. We want the prep phase
to error out rather than the ingestion phase otherwise we can leave the
datastore for that resource in a weird half ingested state and waste
disk space.

Useful ref: http://www.unicode.org/reports/tr44/#General_Category_Values
and https://stackoverflow.com/a/93029.