Check the status as well as the index existance
This came up in my test environment and is unlikely to occur
in real use, however it's worth covering off. When a search or,
for example, is undertaken we check the status index and do
nothing if the index doesn't appear in the status (i.e. it doesn't
have an associated latest version. Therefore, it's not enough to
check that the index exists and we should be checking the status
index too.