Add marine/geography/countries to schema and slightly alter the schema
The change to the schema is similar to the one made when we switched
from having a type key in each term to just using the term name as the
key for the options. Now when using a geo_named_area term instead of:

{
  "category": "country",
  "name": "Finland"
}

you do this:

{
  "country": "Finland"
}

Much simpler to use, more efficient and it's much easier to write a
schema for it.