Sort out relative path references so that we can include geo names enum
This commit has a few related changes which were hard to split up, sozz.
Firstly, the v1.0.0 schema has been moved into a v1.0.0/ directory,
under which it lives along with all the geojson data. In the geojson
directory there are now schemas for each of the named area categories
(country, geography and marine). The readme in the geojson directory has
been updated to be better generally and also includes information about
how to generate the named areas category schemas.

With the addition of the schemas for the named area categories, there
are now references to them in the main schema. These references are
relative and the $ids have been removed to avoid confusion in this
regard. A RefResolver has been added to the core schema loader to handle
these relative schemas correctly.