Allow specifying a list of resource ids to search against
If no resource ids list is passed (or an empty list) then all resources
available to the current user are searched. If a list of resource ids is
provided then each resource id is checked to make sure the current user
can access it and then only those that they can are searched.

We will need to confirm that the access checks are fast enough with lots
of data.