Fairly big refactor after eevee v1.2.0 release which removes search code
Searching now needs to be setup and run from outside of eevee which is
better really as it gives us flexibility on how we run searches.

Because of this change we are now able to sensibly run the multisearch
queries through the multisearch elasticsearch endpoint. This has nothing
to do with the same name-ness, this is caused by a URL length issue.
Because we're querying many indexes and specifying them too, we can
create really long queries. Elasticsearch does have a setting to control
the limit of URL length but in order to completely avoid it being an
issue (and to avoid issues on the NHM internal network too, argh) it's
easier just to use a multisearch with a single search in it as it allows
us to specify the indexes in the body and thus there are no issues!

Note that this commit contains a breaking change in that the
IVersionedDatastore plugin's datastore_modify_result method now takes an
elasticsearch result object rather than an eevee SearchResults object.