Add slug creation and storage in redis
Every query gets a slug. The multisearch action will return a slug with
each response or the create slug action can be used to create a slug
without running a query. The resolve slug action can be used to retrieve
the query, query version, resource ids and version associated with a
slug. The slugs can either be "pretty" and a combination of two
adjectives and an animal or a uuid.