Only allow the properties that we specify
This is useful for ensuring users run the queries they expect to run,
but also with query slugs coming we only want to store queries that have
the correct structure. If we allow additional properties we could end up
with loads of queries with random keys and values that aren't relevant
to the query itself. This isn't great and also will result in different
slugs being generated based on non-query altering parts of the query.