Add basic download support
This supports csvs, tsvs and jsonl outputs. Users can also choose
whether to get the data from each resource in the query in separate
files or in one uber file (this is all put in a zip for the user).
Additionally, users can also specify that they only want fields with
values in them.

Currently the email that is sent to the user is silly and needs to be
given some proper words. Also, no stats are stored.

DwC-A to come too.