Use an or instead of many queries
It should be more efficient to query many indexes at once with an or
on all the fields in the group than run a query for each field in the
group on just the indexes it appears in. This is because we get the
short circuiting powers of or this way whereas by short circuiting
after the queries have already been run we're not short circuiting at
all. Basically this way is faster.