Order the groups in the guess fields action using the number of hits
Prior to this commit the groups were ordered by count then
alphabetically by group name. Now they are ordered by count and then the
total number of records the fields groups appear in. This is more useful
for users as it will ensure the fields shown first are the ones with the
most values in them.

As part of this change a "records" key is also now returned in the
response to this action.