Fix sorting on single specimens
Firstly, the alphabetical by default needs to be brought back to ensure
there is some basic stable order of the fields we get back from the
top_groups function otherwise it could cause some subtly different
results (if you take the first 5 elements from the top_groups return and
then use it, there's no guarantee you'll get the same 5).

Additionally, now if there is no stored order for the fields from the
ingestion we sort them in the same way as we sort the fields if there
are many resources. This is better than just doing alphabetical.