Remove config usage for slug transfer between requests
This doesn't work in multithreaded workflows as the config object is
global and therefore slugs can get mis-assigned very easily. Example:

1. User A requests /search/beans and config['search_slug'] is set to
   'beans'. A redirect is returned to send them to the search base page.
2. User B requests /search/goats and config['search_slug'] is set to
   'goats' - this overrides the 'beans' slug value and it is lost. A
   redirect is returned to send them to the search base page.
3. User A requests /search after getting the redirect back. When the
   page is rendered the config['search_slug'] value is 'goats' (due to
   user B's request) and therefore they get back the search associated
   with the slug 'goats'.
4. User B requests /search after getting the redirect back. When the
   page is rendered the config['search_slug'] value is empty (due to
   user A's second request) and therefore they get back an empty search
   page.

The only option available to us if we want to maintain this way of
passing the slug _between_ requests is to put it in the session data or
the database/another global storage location. The session makes the most
sense at first, however this wouldn't work either as the problem
remains - all using the session would resolve is that this wouldn't be a
problem between user A and user B anymore, just a problem for each user
on their own.

The easiest thing to do therefore is to not bother and just return the
page with the slug in the URL. If the page then has some javascript on
it to remove the slug from the URL then that can happen after it loads.