Only display fields from resources that are in the search results
Because we use an or with existance checks on many fields on many
resources at the same time, we have the potential to return a list of
groups where the group is only represented in one of the resources which
is bad. By adding a quick filter to make sure we're only working on the
resources we know appear in the search, we avoid this problem.