Add a download handling route for dev
If debug mode is on in the ckan config, this route is registered so that
we can serve the download files in the dev environment when we might not
be running apache in front of the ckan server. In the future we may wish
to use this route to collect stats on how many times a download has been
requested after creation and use X-Sendfile to do this efficiently.