Only write fields that exist
This may seem like a pointless exercise as surely if the field count is
0 for a field then it won't appear in any of the data dicts - however,
because the data returned from elasticsearch is the source dict that was
uploaded to it, it could contain nulls and empty string values. The
calculate_field_counts function that generates the field counts does it
by using exist queries and because these count indexed values it skips
nulls and empty strings which are ignored during indexing.