Catch validation errors during autoslug creation
It's possible to start up ckan without any resources and therefore
when the autoslugs are created it's possible to get validation errors
because the requested resource ids don't exist, i.e. one of these:

ValidationError: {'message': u"The requested resources aren't accessible to this user"}

This is fixed by just catching these errors.