Remove pandas dependency as rewrite affected section
Soooooo. The way this was done in pandas was much nicer than the
new implementation but the inclusion of pandas for this one use
isn't worth the downside: 25 minute docker builds!

The docker image in this repo is based on the ckan-dev docker
image which in turn is based on alpine linux. When installing
pandas (and therefore numpy) on a normal libc based linux system
(e.g. an ubuntu based docker image) you get a pre-compiled wheel
and the install is very fast. Alpine linux doesn't use libc and
therefore when pandas and numpy are installed the source is
downloaded and compiled. This takes a really long time cause it's
an enormously complicated library - usually somewhere in the
region of 25-30 mins.

The only way (currently) to get round this and keep pandas would
be to replace the ckan-dev docker base with something else, like
a homebrewed ckan image based on ubuntu. Given the section of code
using pandas isn't critically dependant on pandas (it could be
rewritten in a few different ways, including the way it's done in
this commit), it makes more sense at this point to remove it than
change the docker setup. However, there's no reason we couldn't
change the docker setup in the future and bring back the pandas
dependency.