fix: remove cached package from context
when core ckan checks the package_show auth, the package it checks is cached in the context object and retrieved next time the auth is called, regardless of whether the ID matches. if the auth is called multiple times in a row, the first package checked will be used for all subsequent calls.