feat: apply compatibility changes to work with Splitgill vNext
BREAKING CHANGE: apply compatibility changes to work with Splitgill vNext

This has loads of stuff in it. The major changes are to ingest and sync as related to Splitgill. Additionally, IDs no longer have to be integers, they are strings. There are then a lot of refactoring elements to this commit which cover querying, actions, queued tasks, and a number of other more minor areas. The download code is mostly left as is with a few minor alterations.