build: update pre-commit and pyproject
This updates all of the pre-commit hooks and swaps black for ruff. This is essentially what has already been done on the main branches of all our other extensions but docformatter was broke so I was forced to change the pre-commit and figured I might as well just update it entirely to meet current standard.