feat: upgrade deps
Splitgill is now at v3.0.1 and stable. The ckanext-query-dois commit sha used is the merge into the dev branch of the josh/preps branch so this changes nothing and just pins it at the merge commit. The ckanext-attribution update is just a point release update to be in line with latest.