feat: reimplement vds_multi_direct (datastore_search_raw)
Re-added this action back in with a few changes, most notably that it has a new name and it can only be used by admin users. Admin user access is specified to avoid the action being abused if it was public. For other extensions wishing to use this action, highly recommend just passing ignore_auth in your context.