fix: use max resource version if no requested version specified
Multisearch DOIs created before v6 do not have a requested version, so they were defaulting to "None" which just shows the most recent version of the resource.

This "fixes" this issue but is NOT fully backwards compatible. It has been tested on the NHM's dataset and found to work, but there is the possibility that on other implementations the newest version in the list will not resolve to the correct version for other resources. This has not been marked as a breaking change because it should not break anything more than v6 already will have.

Closes: #187