fix: ignore resources with no database
Temporary fix. At some point this should be fixed by passing the "ignore_unavailable" param into the splitgill search (NaturalHistoryMuseum/splitgill#38), but for now we can at least catch the errors and return some resuts for the resources that do exist.

It really shouldn't get to this point anyway, so this is a last resort for resources that for whatever reason have made it through the previous validation steps.

Closes: #193