fix: use get_action rather than check_access
check_access does not check if the resource exists, so check_resource_id was returning True for resources that do not and have never existed. get_action also checks if the user has access, so we can just replace it.

We should also be using the toolkit exceptions rather than importing directly from ckan.logic.