fix: fix summing logic to account for arc indices
If records in a resource are spread over multiple indices, the agg on _index will produce multiple results for a single resource which under the previous logic was not accounted for. This resulted in a counts dict which only contained the count for a resource from the last index iterated over containing data from that resource. This fixes this issue.

Closes: #208